+++ lib/constants/app_info.dart
@@ -1,8 +1,8 @@
 /// App version and info constants
 /// Update version here only - all other files will reference this
 class AppInfo {
-  static const String version = '2.1.6';
-  static const String buildNumber = '44';
+  static const String version = '2.1.7';
+  static const String buildNumber = '45';
   static const String fullVersion = '$version+$buildNumber';